Rename, clean up examples, fix benchmarks
