+++ newton/examples/__init__.py
@@ -187,12 +187,12 @@ def main():
         "cloth_style3d": "newton.examples.cloth.example_cloth_style3d",
         "mpm_granular": "newton.examples.mpm.example_mpm_granular",
         "mpm_anymal": "newton.examples.mpm.example_mpm_anymal",
-        "robot_anymal_c_walk": "newton.examples.robot.example_anymal_c_walk",
-        "robot_anymal_d": "newton.examples.robot.example_anymal_d",
-        "robot_cartpole": "newton.examples.robot.example_cartpole",
-        "robot_g1": "newton.examples.robot.example_g1",
-        "robot_h1": "newton.examples.robot.example_h1",
-        "robot_humanoid": "newton.examples.robot.example_humanoid",
+        "robot_anymal_c_walk": "newton.examples.robot.example_robot_anymal_c_walk",
+        "robot_anymal_d": "newton.examples.robot.example_robot_anymal_d",
+        "robot_cartpole": "newton.examples.robot.example_robot_cartpole",
+        "robot_g1": "newton.examples.robot.example_robot_g1",
+        "robot_h1": "newton.examples.robot.example_robot_h1",
+        "robot_humanoid": "newton.examples.robot.example_robot_humanoid",
         "selection_articulations": "newton.examples.selection.example_selection_articulations",
         "selection_cartpole": "newton.examples.selection.example_selection_cartpole",
         "selection_materials": "newton.examples.selection.example_selection_materials",
